// frontend/src/types/Api-Item/Ujian/JadwalUjianApiItem.ts
export type JadwalUjianApiItem = {
  id: number;
  id_ujian: number;
  id_guru: number;
  id_pengawas: number;
  nama_ujian: string;
  pengawas_ujian: string;
  tgl_ujian: string;
  tanggal_ujian: string;
  waktu_mulai: string;
  waktu_selesai: string;
  sesi_ujian: number;
  ruang_ujian: string;
  id_ruang: number;
  status_ujian: string;
  started: number;
  tingkat_kelas: number;
  tingkat_kelas_id: number;
  nama_kelas: string;
  pembuat_username: string;
  pengawas_username: string;
};